Fixed a typo in the Mojolicious::Lite POD
I think a mistake exist in your POD. 

Dunno if the README.md is auto-generated from POD, but I guess so, so I didn't corrected it. 

If it isn't a typo, my pull request should be turned to a bug because I can't manage to instantiate it this way with `Mojolicious::Lite`.